end of day commit. Progress in desc
PRoblems to solve:

with cookies now being sent from the client once being saved, i cannot access them from the proxy server. Attempted to depoy to see if working in production but there are issues having the frontend access the backend.

the deployed backened is workign fine on postman - now using auth-token headers instead of bearr tokens.

Think the vercel issue is do do with server side rendering and perhaps it not getting the data it requires for rendering.

auth-token headers have been hardcoded in now so that the deployment can work before solving the cookie access issue